mentored-backend/requirements.txt
aiohttp==3.7.4.post0
Django==4.1.4
djangorestframework_simplejwt==5.0.0
jsonschema==3.2.0
kubernetes==24.2.0
matplotlib==3.5.1
netdata==0.2.1
pandas==1.4.1
Pygments==2.12.0
pysaml2==7.4.1
PyYAML==6.0
rest_framework_simplejwt==0.0.1
six==1.14.0
SQLAlchemy==2.0.5.post1
tqdm==4.63.0
uWSGI==2.0.21
